requirements.txt
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-split-settings==0.3.0
Pillow==6.1.0
